move paths pkg; remove dupe isStdLib func
